feat:Criação do CRUD para a estante de livros e adição de paginação e filtros.

+++ src/main/java/com/acirio/virtual_bookshelf/model/BookModel.java
@@ -29,7 +29,7 @@ public class BookModel {
 
     private String genre;
 
-    private Long numberOfPages;
+    private int numberOfPages;
 
     private String cover; //Endereço da imagem da capa do livro
